initial

+++ app/Http/Controllers/QuoteController.php
@@ -8,6 +8,7 @@ use App\Models\Quote;
 use App\Models\Group;
 use Validator;
 use App;
+use Illuminate\Support\Facades\Auth;
 
 
 
@@ -20,14 +21,13 @@ class QuoteController extends Controller
             $quotes = Quote::where('group', request('group', 'none'))->orderBy( (request('order') == 'popularity'?'views':'id') , 'desc')->get();
         }
         return view('quotes.index', [
-            "lang"=>App::getLocale(),
             "quotes" => $quotes,
             "groups" => Group::all()
         ]);
     }
 
-    public function show(Request $request, $lang, $uuid) {
-        if(!$quote = Quote::firstWhere('uuid', $uuid)) {
+    public function show(Request $request, $hashid) {
+        if(!$quote = Quote::firstWhere('id', decodeId($hashid))) {
             abort(404);
         }
         $quote->views +=1;
@@ -42,49 +42,82 @@ class QuoteController extends Controller
     public function store(Request $request) {
         $validator = Validator::make($request->all(), [
             'group' => 'required',
-            'author' => 'required|string|max:255',
-            'content' => 'required|string',
+            'author' => 'required|string|max:30',
+            'content' => 'required|string|max:150',
         ]);
         if ($validator->fails()) {
             return back()->withInput()->withErrors($validator);
+        }
+
+        if ((!$group = Group::firstWhere('uuid', request('group'))) && request('group')!=="none") {
+            abort('400');
         }
 
         $quote = Quote::create([
-            'uuid' => Str::orderedUuid(),
             'group' => request('group'),
             'author' => request('author'),
             'content' => request('content'),
+            'user' => Auth::user()->hashid,
+            'show' => $group?$group->canUserPost():true
         ]);
-        return redirect()->route('quote.show', ['lang'=>App::getLocale(), 'uuid'=>$quote->uuid]);
+        return redirect()->route('quote.show', ['hashid'=>$quote->hashid])
+            ->with('status', $quote->show?'success':'warn')
+            ->with('content', $quote->show?'Citation ajoutée et visible':'Citation ajoutée, en attente de validation');
     }
 
-    public function edit(Request $request, $lang, $uuid) {
-        return view('quotes.edit', ["groups"=>Group::all(), "quote"=>Quote::firstWhere('uuid', $uuid)]);
+    public function edit(Request $request, $hashid) {
+        if(!$quote = Quote::firstWhere('id', decodeId($hashid))) {
+            abort('404');
+        }
+
+        if(!$quote->canUserTouch())
+            abort('403');
+
+        return view('quotes.edit', ["groups"=>Group::all(), "quote"=>$quote]);
     }
 
-    public function update(Request $request, $lang, $uuid) {
+    public function update(Request $request, $hashid) {
         $validator = Validator::make($request->all(), [
-            'author' => 'required|string|max:255',
-            'content' => 'required|string',
+            'author' => 'required|string|max:30',
+            'content' => 'required|string|max:150',
         ]);
         if ($validator->fails()) {
             return back()->withInput()->withErrors($validator);
         }
+        if(!$quote = Quote::firstWhere('id', decodeId($hashid))) {
+            abort('404');
+        }
 
-        $quote = Quote::firstWhere('uuid', $uuid);
+        if(!$quote->canUserTouch())
+            abort('403');
+
+        if ((!$group = Group::firstWhere('uuid', $quote->group)) && request('group')!=="none") {
+            abort('400');
+        }
+
         $quote->author = request('author');
         $quote->content = request('content');
+        $quote->show = $group?$group->canUserPost():true;
         $quote->touch();
         $quote->save();
 
-        return redirect()->route('quote.show', ['lang'=>App::getLocale(), 'uuid'=>$quote->uuid]);
+        return redirect()
+            ->route('quote.show', ['hashid'=>$quote->hashid])
+            ->with('status', $quote->show?'success':'warn')
+            ->with('content', $quote->show?'Citation modifiée et visible':'Citation modifiée, en attente de validation');
     }
 
-    public function destroy(Request $request, $lang, $uuid) {
-        $quote = Quote::firstWhere('uuid', $uuid);
+    public function destroy(Request $request, $hashid) {
+        if(!$quote = Quote::firstWhere('id', decodeId($hashid))) {
+            abort('403');
+        }
+
+        if(!$quote->canUserTouch())
+            abort('403');
+
         $quote->delete();
 
-        return redirect()->route('home', ['lang'=>App::getLocale()]);
+        return redirect()->route('home');
     }
